requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.4
cffi==1.14.4
chardet==3.0.4
discord.py==1.6.0
emojis==0.6.0
idna==3.1
multidict==5.1.0
pycodestyle==2.6.0
pycparser==2.20
PyNaCl==1.4.0
python-dotenv==0.15.0
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
yarl==1.6.3
youtube-dl==2021.1.16
youtube-search-python==1.3.6
yt-search==1.1.0.post1
